#1 - 아키텍쳐 변경
