+++ app/src/main/AndroidManifest.xml
@@ -32,10 +32,6 @@
             android:name=".LoginActivity"
             android:exported="true"
             android:theme="@style/Theme.AppCompat">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
         </activity>
 
         <activity
