+++ app/src/main/java/com/demo/desh/MainActivity.kt
@@ -1,11 +1,17 @@
 package com.demo.desh
 
+import android.content.Context
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.activity.compose.setContent
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Column
 import androidx.compose.material3.Surface
-import com.demo.desh.dto.User
-import com.demo.desh.ui.screen.MainActivityScreen
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.tooling.preview.Preview
+import coil.compose.rememberAsyncImagePainter
+import com.demo.desh.model.User
 import com.demo.desh.ui.theme.DeshprojectfeTheme
 
 class MainActivity : AppCompatActivity() {
@@ -21,5 +27,28 @@ class MainActivity : AppCompatActivity() {
                 }
             }
         }
+    }
+}
+
+@Composable
+fun MainActivityScreen(context: Context, user: User) {
+    Text(text = "MainScreen")
+
+    Column {
+        Text(text = user.id!!.toString())
+        Text(text = user.email)
+        Text(text = user.nickname)
+
+        Image(
+            painter = rememberAsyncImagePainter(model = user.profileImageUrl),
+            contentDescription = ""
+        )
+    }
+}
+
+@Preview(showSystemUi = true, showBackground = true)
+@Composable
+private fun MainActivityScreenPreview() {
+    DeshprojectfeTheme {
     }
 }
