+++ app/src/main/java/com/demo/desh/SurveyActivity.kt
@@ -25,7 +25,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
 import com.demo.desh.navigation.NavGraph
 import com.demo.desh.navigation.Screen
-import com.demo.desh.dto.User
+import com.demo.desh.model.User
 import com.demo.desh.ui.theme.DeshprojectfeTheme
 import com.demo.desh.ui.theme.nanum
