+++ app/src/main/java/com/demo/desh/viewModels/LoginViewModel.kt
@@ -1,4 +1,4 @@
-package com.demo.desh.viewmodel
+package com.demo.desh.viewModels
 
 import android.content.Context
 import androidx.lifecycle.ViewModel
